refactoring: Move (from ZM) and refactor vanilla KZG10 utilities
